Move LoadingScreen style
